--- a/package.json
+++ b/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "precommit": "lint-staged",
     "postcommit": "git reset",
-    "prettier": "prettier \"packages/**/{src,test}/*.{ts,tsx}\" --write --print-width 100 --tab-width 4",
+    "prettier": "prettier \"packages/**/{src,test}/*.{js,ts,tsx}\" --write --print-width 120 --tab-width 4",
     "bootstrap": "lerna bootstrap",
     "prepare": "npm run bootstrap",
     "test": "jest",