… packages/benchmark/src/util.js